chore: encapsulate `bpmn` in own module
