+++ lib/bpmn/attach/BoundaryAttachNode.js
@@ -4,7 +4,7 @@ import {
   domify as domify
 } from 'min-dom';
 
-import AppendNode from '../append/AppendNode';
+import AppendNode from '../../common/append/AppendNode';
 
 const OFFSET = 23;
