+++ lib/bpmn/contextPad/ImprovedContextPadProvider.js
@@ -1,5 +1,5 @@
 import { is } from 'bpmn-js/lib/util/ModelUtil';
-import icons from './ContextPadIcons';
+import icons from '../../common/contextPad/ContextPadIcons';
 
 import {
   query as domQuery
