+++ lib/bpmn/contextPad/index.js
@@ -1,11 +1,12 @@
-import ImprovedContextPad from './ImprovedContextPad';
+import ImprovedContextPad from '../../common/contextPad';
 import ImprovedContextPadProvider from './ImprovedContextPadProvider';
 
 export default {
+  __depends__: [
+    ImprovedContextPad
+  ],
   __init__: [
-    'improvedContextPad',
     'improvedContextPadProvider'
   ],
-  improvedContextPad: [ 'type', ImprovedContextPad ],
   improvedContextPadProvider: [ 'type', ImprovedContextPadProvider ]
 };
