+++ lib/bpmn/popupMenu/AlignElementsIcons.js
