+++ lib/common/append/AppendEditorActions.js
