+++ lib/common/append/AppendKeyboardBindings.js
