+++ lib/common/append/AppendNode.js
