+++ lib/common/contextPad/ImprovedContextPad.js
@@ -20,7 +20,7 @@ import {
 
 import { TooltipEntry } from '@bpmn-io/properties-panel';
 
-import { render } from 'preact';
+import { render } from '@bpmn-io/properties-panel/preact';
 
 import { html } from 'htm/preact';
