+++ lib/common/contextPad/index.js
@@ -1,0 +1,8 @@
+import ImprovedContextPad from './ImprovedContextPad';
+
+export default {
+  __init__: [
+    'improvedContextPad'
+  ],
+  improvedContextPad: [ 'type', ImprovedContextPad ]
+};
