+++ lib/common/highContrastCanvas/HighContrastCanvas.js
