+++ lib/common/highContrastCanvas/index.js
