+++ lib/common/popupMenu/ColorPickerPopupProvider.js
