+++ lib/common/popupMenu/HorizontalPopupMenu.js
