+++ lib/common/popupMenu/index.js
@@ -1,14 +1,11 @@
 import HorizontalPopupMenu from './HorizontalPopupMenu';
 import ColorPickerPopupProvider from './ColorPickerPopupProvider';
-import AlignElementsPopupProvider from './AlignElementsPopupProvider';
 
 export default {
   __init__: [
     'horizontalPopupMenu',
     'colorPickerPopupProvider',
-    'alignElementsPopupProvider'
   ],
   horizontalPopupMenu: [ 'type', HorizontalPopupMenu ],
-  colorPickerPopupProvider: [ 'type', ColorPickerPopupProvider ],
-  alignElementsPopupProvider: [ 'type', AlignElementsPopupProvider ]
+  colorPickerPopupProvider: [ 'type', ColorPickerPopupProvider ]
 };
