+++ lib/index.js
@@ -1,4 +1,1 @@
-export { default as ImprovedCanvasModule } from './ImprovedCanvas';
-
-// features
-export { default as HighContrastCanvasModule } from './features/highContrastCanvas';
+export { default as BpmnImprovedCanvasModule } from './bpmn';
