+++ test/spec/bpmn/append/AppendMenuProvider.spec.js
@@ -7,13 +7,13 @@ import {
   query as domQuery,
   queryAll as domQueryAll
 } from 'min-dom';
-import Append from 'lib/features/append';
+import Append from 'lib/bpmn/append';
 
 import {
   CreateAppendAnythingModule
 } from 'bpmn-js-create-append-anything';
 
-import diagramXML from './AppendNode.bpmn';
+import diagramXML from '../../common/append/AppendNode.bpmn';
 
 describe('<AppendMenuProvider>', function() {
