+++ test/spec/bpmn/attach/BoundaryAttachNode.bpmn
