+++ test/spec/common/append/AppendEditorActions.spec.js
@@ -7,7 +7,7 @@ import {
 import {
   query as domQuery
 } from 'min-dom';
-import Append from 'lib/features/append';
+import Append from 'lib/bpmn/append';
 
 import {
   CreateAppendAnythingModule
