+++ test/spec/common/contextPad/ImprovedContextPad.spec.js
@@ -15,7 +15,7 @@ import {
   getBBox
 } from 'diagram-js/lib/util/Elements';
 
-import ImprovedContextPad from 'lib/features/contextPad';
+import ImprovedContextPad from 'lib/bpmn/contextPad';
 
 import ColorPickerModule from 'bpmn-js-color-picker';
 
@@ -103,74 +103,6 @@ describe('<ImprovedContextPad>', function() {
       // then
       expect(dragging.context().prefix).to.eql('connect');
 
-    }));
-
-  });
-
-
-  describe('entries', function() {
-
-    it('single element', inject(function(elementRegistry, contextPad) {
-
-      // given
-      const shape = elementRegistry.get('StartEvent_1');
-
-      // when
-      contextPad.open(shape);
-
-      // then
-      const entries = [ ...domQueryAll('.djs-context-pad .entry') ];
-      expect(entries.length).to.equal(5);
-      expect(entries.map(entry => entry.getAttribute('data-action'))).to.eql([
-        'replace',
-        'set-color',
-        'append.text-annotation',
-        'connect',
-        'delete'
-      ]);
-    }));
-
-
-    it('multiple elements', inject(function(elementRegistry, contextPad) {
-
-      // given
-      const shape1 = elementRegistry.get('StartEvent_1');
-      const shape2 = elementRegistry.get('Task_1');
-
-
-      // when
-      contextPad.open([ shape1, shape2 ]);
-
-      // then
-      const entries = [ ...domQueryAll('.djs-context-pad .entry') ];
-      expect(entries.length).to.equal(3);
-      expect(entries.map(entry => entry.getAttribute('data-action'))).to.eql([
-        'set-color',
-        'align-elements',
-        'delete'
-      ]);
-    }));
-
-
-    it('participants', inject(function(elementRegistry, contextPad) {
-
-      // given
-      const shape = elementRegistry.get('Participant_1');
-
-      // when
-      contextPad.open(shape);
-
-      // then
-      const laneGroup = domQuery('.djs-context-pad .group[data-group="lanes"]');
-      expect(laneGroup).to.exist;
-
-      const entries = [ ...domQueryAll('.djs-context-pad .entry', laneGroup) ];
-      expect(entries.map(entry => entry.getAttribute('data-action'))).to.eql([
-        'lane-insert-above',
-        'lane-divide-two',
-        'lane-divide-three',
-        'lane-insert-below'
-      ]);
     }));
 
   });
